requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
bcrypt
plotly
google-generativeai
openpyxl
python-dotenv
xlsxwriter
google-genai